Version stable V1.3 All API UserAccount, Wish, Watch

+++ MMA-ApplicationLayer/src/main/java/com/bcefit/projet/service/watch/IWatchSeasonService.java
@@ -6,7 +6,7 @@ import com.bcefit.projet.domain.watch.WatchEpisode;
 import java.util.List;
 
 public interface IWatchSeasonService {
-    List<WatchEpisode> createWatchEpisodeBySeasonId(Integer idTv, Integer idSeason, UserAccount userAccount);
+    List<WatchEpisode> createWatchEpisodeBySeasonId(Long idTv, Long idSeason, UserAccount userAccount);
 
-    List<WatchEpisode> deleteWatchEpisodeBySeasonId(Integer idTv, Integer idSeason, UserAccount userAccount);
+    List<WatchEpisode> deleteWatchEpisodeBySeasonId(Long idTv, Long idSeason, UserAccount userAccount);
 }
